test1 y test2 LISTOS

+++ T0/src/runner/runner.c
@@ -3,13 +3,12 @@
 
 
 void handle_process(int n_argument, char* path, int index){
-    printf("####################################\n");
-    printf("ENTRE A HANDLE_PROCESS\n");
+    printf("\nENTRE A HANDLE_PROCESS\n");
     printf("\n");
     // número de argumentos de proceso, path+argumentos, indice del proceso.
     if (n_argument == -1) {
         // WAIT_ALL
-        start_time[index] = (double)clock()/(double)CLOCKS_PER_SEC;
+        start_time[index] = GetTime();
         printf("START_TIME WAIT: %f del proceso %s\n", start_time[index], path);
         arguments = malloc(sizeof(char*) * 2);
         split_string(path, " ", arguments);
@@ -18,17 +17,14 @@ void handle_process(int n_argument, char* path, int index){
         printf("timeout: %d", timeout);
 
         wait_all_childs(timeout);
-        end_time[index] = (double)clock()/(double)CLOCKS_PER_SEC; 
+        end_time[index] = GetTime();
         printf("END_TIME WAIT: %f del proceso %s\n", end_time[index], path);
         // free(arguments);
 
     }
 
-    // cont_childs es el numero de hijos que estan corriendo
     else if (cont_childs <= max_childs) {
         // Iniciamos el proecos externo
-        // char **arguments = split_by_sep(path, " ");
-        // arguments[MAX_SPLIT-1] = NULL;
     
         arguments = malloc(sizeof(char*) * (n_argument + 2));
         split_string(path, " ", arguments);
@@ -41,6 +37,9 @@ void handle_process(int n_argument, char* path, int index){
             args[i-1] = arg;
             printf("Argumento %d: %s \n", i, arg);
         }
+        printf("\n");
+
+        start_time[index] = GetTime();
 
         int pid = fork();
         if (pid == -1) {
@@ -52,18 +51,17 @@ void handle_process(int n_argument, char* path, int index){
             printf("\n");
             // execv(arguments[0], arguments); // 
             execvp(arguments[0], args);
+            actual_status[index] = "STARTED";
             printf("Argumeto: %s, Argumnets: %s\n", arguments[0], *args);
             printf("Error ejecutando proceso %s\n", path);
             perror(execvp);
 
             // free(arguments);
             // free(args);
-            exit(1);
         } else {
             // Parent
             printf("ENTRE A PARENT\n");
             init_output(index, arguments[0], pid);
-            // agregro pid, nombre del programa y start_time
             // free(arguments);
             // free(args);
 
@@ -73,35 +71,19 @@ void handle_process(int n_argument, char* path, int index){
             } else {
                 printf("ENTRE AL LISTADO DE PROCESOS...\n");
                 childs[indice] = pid;
+                // printf("childs pid: %d\n", childs[indice]);
                 cont_childs++;
             }
             printf("SALI DE PARENT\n");
             printf("\n");
 
         }
+
         free(args);
         free(arguments);
-        // } else {
-        //     if (pid == 0) {
-        //         execvp(arguments[0], args);
-        //         perror(execvp);
-        //         exit(1);
-        //     } else {
-        //         init_output(index, arguments[0], pid); // agregro pid, nombre del programa y start_time
-        //         int indice = available_space();
-        //         if (indice == -1) {
-        //             printf("No hay espacio disponible\n");
-        //         } else {
-        //             childs[indice] = pid;
-        //             cont_childs++;
-        //         }
-        //         printf("\n");
-        //     }
-        // }
     }
 
     else {
-        printf("QUE PASO??\n");
         wait_child();
         return handle_process(n_argument, path, index);
     }
@@ -117,8 +99,7 @@ void wait_child() {
             int exit_status = WEXITSTATUS(status);
             p_status[index] = exit_status;
             actual_status[index] = "FINISHED";
-            // end_time[index] = time();
-            end_time[index]= (double)clock()/(double)CLOCKS_PER_SEC;
+            end_time[index] = GetTime();
         }
         
     }
@@ -129,7 +110,7 @@ void wait_child() {
 void wait_all_childs(int timeout) {
     printf("Waiting for all childs to finish, max time: %d seconds.\n", timeout);
     int status;
-    double startTime = (double)clock() / CLOCKS_PER_SEC;
+    double startTime = GetTime();
     double elapsedTime = 0;
 
     while (cont_childs > 0 && elapsedTime < timeout) {
@@ -144,13 +125,14 @@ void wait_all_childs(int timeout) {
                         p_status[i] = WTERMSIG(status);
                     }
                     actual_status[i] = "FINISHED";
-                    end_time[i] = (double)clock() / CLOCKS_PER_SEC;
+                    end_time[i] = GetTime();
                     childs[i] = -1;
                     cont_childs--;
                 }
             }
         }
-        elapsedTime = (double)clock() / CLOCKS_PER_SEC - startTime;
+
+        elapsedTime = GetTime() - startTime;
         sleep(1); // Pausa breve para evitar sobrecargar el CPU ESTO ME SUENA RARO...
     }
 
@@ -169,19 +151,45 @@ void wait_all_childs(int timeout) {
 }
 
 // SEÑALES
-void handle_sigint(int signal) {
-    printf("SE TERMINO EL TIEMPO MAX\n");
-    sleep(10);
-    // STATUS = 2
+
+void handle_sigalrm(int signal) {
+    printf("SIGALRM\n");
     for (int i = 0; i < max_childs; i++) {
-      if (childs[i] != -1) {
-        kill(childs[i], SIGTERM);
-        // tengo que ended_output
-        // end_output(childs[i], 2);
-        // childs[i] = -1; // esto tampoco debería ser así...
-      }
-  }
+        if (childs[i] != -1) {
+            kill(childs[i], SIGINT);
+        }
+    }
+
+    sleep(10);
+    for (int i = 0; i < max_childs; i++) {
+        if (childs[i] != -1) {
+            int status;
+            pid_t result = waitpid(childs[i], &status, WNOHANG);
+            if ( result == 0 ) {
+                kill(childs[i], SIGTERM);
+            }
+            end_time[i] = GetTime();
+            actual_status[i] = "FINISHED";
+            p_status[i] = WTERMSIG(status); // SIEMPRE TIENE QUE SER 2, VERIFICAR
+            printf("MAX TIMEPO ALCANZADO\nstatus: %d\n, proceso: %s\n", p_status[i], p_path[i]);
+        }
+    }
 }
+
+// void handle_sigint(int signal) {
+//     printf("SE TERMINO EL TIEMPO MAX\n");
+//     sleep(10);
+//     // STATUS = 2
+//     for (int i = 0; i < max_childs; i++) {
+//       if (childs[i] != -1) {
+//         kill(childs[i], SIGTERM);
+//         // tengo que ended_output
+//         end_output(childs[i], 2);
+//         actual_status[i] = "FINISHED";
+//         // childs[i] = -1; // esto tampoco debería ser así...
+//       }
+//   }
+// }
 
 void handle_sigtstp(int signal){
     // time_reached = true;
@@ -191,9 +199,8 @@ void handle_sigtstp(int signal){
         if (childs[i] != -1) {
             kill(childs[i], SIGTERM);
             childs[i] = -1;
-            // double final_time = GetTime();
-            double final_time = (double)clock()/(double)CLOCKS_PER_SEC;
-            end_time[i] = final_time;
+            end_time[i] = GetTime();
+            actual_status[i] = "FINISHED";
         }
     }
     // Esperamos 10 segundos
@@ -207,6 +214,7 @@ void handle_sigtstp(int signal){
             end_output(childs[i], 3); // 
             kill(childs[i], SIGKILL);
             childs[i] = -1;
+            actual_status[i] = "FINISHED";
         }
     }
     generete_output();
@@ -217,22 +225,15 @@ void handle_sigtstp(int signal){
 
 // GENERAR OUTPUT
 void init_output(int index, char* path, int pid) {
-
-  char* program_name = malloc((strlen(path) + 1) * sizeof(char));
-  strcpy(program_name, path);
-  p_path[index] = program_name;
-  p_pids[index] = pid;
-  start_time[index] = (double)clock()/(double)CLOCKS_PER_SEC;
-  actual_status[index] = "STARTED";
-//   double time = GetTime();
-//   start_time[index] = time;
-  printf("START TIME = %f del proceso %s\n", start_time[index], path);
-
+    printf("INIT_OUTPUT\n");
+    char* program_name = malloc((strlen(path) + 1) * sizeof(char));
+    strcpy(program_name, path);
+    p_path[index] = program_name;
+    p_pids[index] = pid;
 }
 
 void end_output(int pid, int status) {
-    // double time = GetTime();
-    // double time = time()
+    // time_t time = time()
     int index;
     for (int i = 0; i < N_process; i++) {
         if (p_pids[i] == pid) {
@@ -240,32 +241,31 @@ void end_output(int pid, int status) {
         break;
         }
     }
-    end_time[index] = (double)clock()/(double)CLOCKS_PER_SEC;
+    end_time[index] = GetTime();
     p_status[index] = status;
     actual_status[index] = "FINISHED";
     printf("TIEMPO FINAL: %f proceos: %s\n", end_time[index], p_path[index]);
 }
 
 void generete_output() {
-    // creo que puedo estar generando el output dos veces...
     printf("\nLLEGUE A GENERAR EL OUTPUT\n");
     for (int i = 0; i < N_process; i++) {
         if (p_path[i] !=  NULL) {
             if (strcmp(actual_status[i], "FINISHED") != 0) {
-                end_time[i] = (double)clock()/(double)CLOCKS_PER_SEC;
-                p_status[i] = 100; // hay que ponerle un status.
-                actual_status[i] = "FINISHED";
-                printf("TERMINE EN GENERETE_OUTPUT con end_time: %f: \n", end_time[i]);
+                printf("\nEl procesos termina en generar output\n");
+                int status;
+                waitpid(childs[i], &status, 0);
+                end_time[i] = GetTime();
+                p_status[i] = WEXITSTATUS(status);
+
             }
-            printf("TIMEPO inicio %f proceso %s\n",start_time[i], p_path[i]);
-            double total_time = end_time[i] - start_time[i];
-            // int status = p_status[i];
-            // char* command = p_path[i];
-            fprintf(output_file, "%s,%f,%d\n", p_path[i], total_time, p_status[i]); 
-            printf("GENERAR OUTPUT FIN\n########################\n");
-            printf("\n");
+            printf("proceso %s\nstatus %d\n", p_path[i], p_status[i]);
+            int total_time = (int)(end_time[i] - start_time[i]);
+            fprintf(output_file, "%s,%d,%d", p_path[i], total_time, p_status[i]); 
         }
     }
+    printf("GENERAR OUTPUT FIN\n");
+    printf("\n");
 }
 
 
@@ -295,12 +295,12 @@ void split_string(char* string, char* delimiter, char** result) {
 }
 
 // TIEMPOS
-// double GetTime() {
-//     struct timeval t;
-//     int rc = gettimeofday(&t, NULL);
-//     assert(rc == 0);
-//     return (double)t.tv_sec + (double)t.tv_usec/1e6;
-// }
+double GetTime() {
+    struct timeval t;
+    int rc = gettimeofday(&t, NULL);
+    assert(rc == 0);
+    return (double)t.tv_sec + (double)t.tv_usec/1e6;
+}
 
 
 // LIBERAR MEMORIA
